requirements.txt
hydra-core==1.3.2
lightning==2.0.2
torchdata==0.6.1
torch
wandb
tensorboard